requirements.txt
ccxt
python-dotenv
fastapi
redis
types-redis
uvicorn
